+++ ddtrace/contrib/internal/django/response.py
@@ -25,9 +25,9 @@ from ddtrace.ext import http
 from ddtrace.internal import core
 from ddtrace.internal._exceptions import BlockingException
 from ddtrace.internal._exceptions import find_exception
+from ddtrace.internal._identity import maybe_refresh_identity
 from ddtrace.internal.constants import COMPONENT
 from ddtrace.internal.logger import get_logger
-from ddtrace.internal.runtime import maybe_refresh_identity
 from ddtrace.internal.schema import schematize_url_operation
 from ddtrace.internal.schema.span_attribute_schema import SpanDirection
 from ddtrace.internal.settings.integration import IntegrationConfig
