+++ ddtrace/contrib/internal/falcon/middleware.py
@@ -3,7 +3,7 @@ import sys
 from ddtrace import config
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import core
-from ddtrace.internal.runtime import maybe_refresh_identity
+from ddtrace.internal._identity import maybe_refresh_identity
 from ddtrace.internal.schema import SpanDirection
 from ddtrace.internal.schema import schematize_service_name
 from ddtrace.internal.schema import schematize_url_operation
