+++ ddtrace/contrib/internal/molten/patch.py
@@ -10,7 +10,7 @@ from ddtrace.contrib import trace_utils
 from ddtrace.contrib.internal.trace_utils import unwrap as _u
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import core
-from ddtrace.internal.runtime import maybe_refresh_identity
+from ddtrace.internal._identity import maybe_refresh_identity
 from ddtrace.internal.schema import schematize_service_name
 from ddtrace.internal.schema import schematize_url_operation
 from ddtrace.internal.schema.span_attribute_schema import SpanDirection
