+++ ddtrace/contrib/internal/sanic/patch.py
@@ -9,8 +9,8 @@ from ddtrace._trace.pin import Pin
 from ddtrace.contrib import trace_utils
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import core
+from ddtrace.internal._identity import maybe_refresh_identity
 from ddtrace.internal.logger import get_logger
-from ddtrace.internal.runtime import maybe_refresh_identity
 from ddtrace.internal.schema import schematize_service_name
 from ddtrace.internal.schema import schematize_url_operation
 from ddtrace.internal.schema.span_attribute_schema import SpanDirection
